Add defendant_searcher.rb service object
Tweak controllers
