+++ app/services/defendant_searcher.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+class DefendantSearcher
+  def self.call(*args, &block)
+    new(*args, &block).call
+  end
+
+  def initialize(search)
+    @search = search
+  end
+
+  def call
+    @search.execute.first
+  end
+end
